Change Edn's usize/isize to u64/i64.

+++ examples/complex_struct_deserialization.rs
@@ -3,7 +3,7 @@ use edn_rs::{Deserialize, Edn, EdnError};
 #[derive(Debug, PartialEq)]
 struct Another {
     name: String,
-    age: usize,
+    age: u64,
     cool: bool,
 }
